final_gemini_chatbot/requirements.txt
streamlit==1.37.1
python-dotenv>=1.0.1
google-generativeai>=0.7.2
langchain>=0.2.12
langchain-community>=0.2.10
langchain-google-genai>=1.0.7
faiss-cpu>=1.8.0.post1
pypdf>=4.2.0
